+++ src/App.js
@@ -8,8 +8,8 @@ import { BrowserRouter as Router, Route } from 'react-router-dom';
 import NavBar from './Components/NavBar';
 import Politicians from './Containers/Politicians';
 import PromiseForm from './Containers/PromiseForm';
-import SignUpForm from './Containers/SignUpForm';
-import LoginForm from './Containers/LoginForm';
+import SignUpContainer from './Containers/SignUpContainer';
+import LoginContainer from './Containers/LoginContainer';
 
 injectTapEventPlugin();
 
@@ -22,8 +22,8 @@ class App extends Component {
             <NavBar />
             <Route exact path="/" component={Politicians} />
             <Route path="/add-promise" component={PromiseForm} />
-            <Route path="/sign-up" component={SignUpForm} />
-            <Route path="/login" component={LoginForm} />
+            <Route path="/sign-up" component={SignUpContainer} />
+            <Route path="/login" component={LoginContainer} />
           </div>
         </Router>
       </MuiThemeProvider>
